@@ -26,7 +26,7 @@ class ErrorPresenter extends \Nette\Application\UI\Presenter
 
 		} elseif ($exception instanceof NA\BadRequestException) {
 			$code = $exception->getCode();
-			$this->setView(in_array($code, array(403, 404, 405, 410, 500)) ? $code : '4xx'); // load template 403.latte or 404.latte or ... 4xx.latte
+			$this->setView(in_array($code, [403, 404, 405, 410, 500]) ? $code : '4xx'); // load template 403.latte or 404.latte or ... 4xx.latte
 			Debugger::log("HTTP code $code", 'access'); // log to access.log
 
 		} else {
@@ -36,4 +36,3 @@ class ErrorPresenter extends \Nette\Application\UI\Presenter
 	}
 
 }
-?>
