@@ -1,5 +1,5 @@
 <?php
 
-$container = include __DIR__ . '/../app/bootstrap.php';
+$container = include __DIR__ . "/../app/bootstrap.php";
 
-$container->getService('application')->run();
+$container->getService("application")->run();
