@@ -128,7 +128,7 @@ Nette.validateRule = function(elem, op, arg) {
 		op = op.substr(1);
 	}
 	op = op.replace('::', '_');
-  op = op.replace('\\', '');
+	op = op.replace('\\', '');
 	return Nette.validators[op] ? Nette.validators[op](elem, arg, val) : null;
 };
 
